// ui/src/generated/RuntimeInstances.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { WorkspaceInstance } from "./WorkspaceInstance";

export type RuntimeInstances = { instances: Array<WorkspaceInstance>, default_workspace_instance_id: string | null, total: number, next_after_instance_id: string | null, };