// ui/src/generated/RuntimeEffectivePolicy.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RuntimePolicy } from "./RuntimePolicy";
import type { RuntimePolicyOverrides } from "./RuntimePolicyOverrides";

export type RuntimeEffectivePolicy = { value: RuntimePolicy, app: RuntimePolicyOverrides, project: RuntimePolicyOverrides, instance: RuntimePolicyOverrides, };