// ui/src/generated/CheckpointList.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CheckpointEntry } from "./CheckpointEntry";

export type CheckpointList = { entries: Array<CheckpointEntry>, next: string | null, native_capture_available: boolean, notice: string | null, };